third_party/blink/web_tests/editing/selection/select_all/select_all_contenteditable_br.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<div id="div" contenteditable><br>foo</div>
<script>
test(function() {
    var div = document.getElementById('div');
    div.focus();
    assert_true(document.queryCommandEnabled('SelectAll'));
}, 'SelectAll is enabled if the contenteditable contains BR (at the beginning) and other content');
</script>
